feat: Split Python Frameworks
