+++ helm/charts/determined/values.yaml
@@ -1,4 +1,4 @@
-# The image registry to be used to pull the Master image.
+4# The image registry to be used to pull the Master image.
 # Determined OSS edition uses the determinedai repository in DockerHub.
 imageRegistry: determinedai
 # HPE Machine Learning Development Environment (MLDE), Determined Enterprise Edition, uses the HPE MSC as the image registry
@@ -31,8 +31,8 @@ defaultImages:
   kubeSchedulerPreemption: "determinedai/kube-scheduler:0.17.0"
 
   # default images for CPU and GPU environments
-  cpuImage: "determinedai/pytorch-tensorflow-cpu-dev:8b3bea3"
-  gpuImage: "determinedai/pytorch-tensorflow-cuda-dev:8b3bea3"
+  cpuImage: "determinedai/pytorch-ngc-dev:8b3bea3"
+  gpuImage: "determinedai/pytorch-ngc-dev:8b3bea3"
 
 # Install Determined enterprise edition.
 enterpriseEdition: false
